Split profile marker migrations
