+++ service-worker.js
@@ -107,6 +107,7 @@ const APP_SHELL = [
   '/js/import-drop-zone-runtime.js',
   '/js/ai-verdict-engine-runtime.js',
   '/js/ai-verdict-engine.js',
+  '/js/profile-marker-migrations.js',
   '/js/profile.js',
   '/js/profile-runtime.js',
   '/js/data.js',
