+++ tsconfig.checkjs.json
@@ -183,6 +183,7 @@
     "js/pdf-import.js",
     "js/pdfjs-loader.js",
     "js/pii.js",
+    "js/profile-marker-migrations.js",
     "js/profile-share.js",
     "js/profile.js",
     "js/profile-runtime.js",
